bug(helpers): moving code to helpers file[finishes #167181853]
